+++ src/extensions/mvs/components/annotation-label/representation.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2023-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Adam Midlik <midlik@gmail.com>
  */
@@ -9,6 +9,7 @@ import { Representation, RepresentationContext, RepresentationParamsGetter } fro
 import { ComplexRepresentation, StructureRepresentation, StructureRepresentationProvider, StructureRepresentationStateBuilder } from '../../../../mol-repr/structure/representation';
 import { MarkerAction } from '../../../../mol-util/marker-action';
 import { ParamDefinition as PD } from '../../../../mol-util/param-definition';
+import { isMVSStructure } from '../is-mvs-model-prop';
 import { MVSAnnotationLabelTextParams, MVSAnnotationLabelTextVisual } from './visual';
 
 
@@ -45,5 +46,5 @@ export const MVSAnnotationLabelRepresentationProvider = StructureRepresentationP
     defaultValues: PD.getDefaultValues(MVSAnnotationLabelParams),
     defaultColorTheme: { name: 'uniform' }, // this ain't workin
     defaultSizeTheme: { name: 'physical' },
-    isApplicable: (structure: Structure) => structure.elementCount > 0,
+    isApplicable: (structure: Structure) => structure.elementCount > 0 && isMVSStructure(structure),
 });